refactor(navbar): decompose Navbar into focused sub-components (#331)
- Extract NAV_ITEMS data to nav/NavLinks.ts
- Extract NavItemLink reusable component for desktop/mobile variants
- Extract DesktopNav with admin dropdown
- Extract MobileNav with auth buttons
- Simplify Navbar.tsx from 400 to 101 lines (orchestrator)
- Remove old MobileMenu.tsx (dead code, 185 lines)
- Update barrel exports
