+++ frontend/src/components/layout/index.ts
@@ -5,6 +5,8 @@
  */
 export { AppLayout, default as AppLayoutDefault } from './AppLayout';
 export { Navbar } from './Navbar';
-export { MobileMenu } from './MobileMenu';
 export { LanguageSelector } from './LanguageSelector';
 export { UserMenu } from './UserMenu';
+export { DesktopNav } from './nav/DesktopNav';
+export { MobileNav } from './nav/MobileNav';
+export { NavItemLink } from './nav/NavItemLink';
